Fix the following bugs:  1377537 1377448 1374720
1) argv[0] was being corrupted by strsep(3), causing ps to show garbage
instead of the process name.  Fixed this by strdup()ing argv[0] before
hacking at it with strsep(3)

2) Linux apparently sends POLLHUP when the remote end has closed, instead of
POLLHUP|POLLIN like NetBSD does.  This means that unless we look for
POLLHUP in the scan, we never get the data, causing dsh to truncate
output from the child rsh/ssh.

3) Add a quick routine to strip the trailing whitespace from the
groupname when parsing the clusterfile.  I think I still need to do this
for the lumpname too.